requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
langchain-mistralai
langchain-chroma
langchain-huggingface
langchain-text-splitters
langchain-community
langgraph
requests
pysqlite3-binary